+++ R/variable_importance.R
@@ -1,5 +1,9 @@
 #' Feature Importance Calculated as Loss from Feature Dropout
 #'
+#' This function is set deprecated. It is suggested to use \code{\link[ingredients]{feature_importance}} instead.
+#' Find information how to use these functions here: \url{https://pbiecek.github.io/PM_VEE/variableImportance.html}.
+#'
+#'
 #' @param explainer a model to be explained, preprocessed by the 'explain' function
 #' @param loss_function a function thet will be used to assess variable importance
 #' @param ... other parameters
@@ -47,6 +51,12 @@ variable_importance <- function(explainer,
                               ...,
                               type = "raw",
                               n_sample = 1000) {
+  # Deprecated, but print the message only once
+  if (!exists("message_variable_importance", envir = .DALEX.env)) {
+    .DALEX.env$message_variable_importance = TRUE
+    .Deprecated("ingredients::feature_importance()", package = "ingredients", msg = "Please note that 'variable_importance()' is now deprecated, it is better to use 'ingredients::feature_importance()' instead.\nFind examples and detailed introduction at: https://pbiecek.github.io/PM_VEE/variableImportance.html")
+  }
+
   if (!("explainer" %in% class(explainer))) stop("The variable_importance() function requires an object created with explain() function.")
   if (is.null(explainer$data)) stop("The variable_importance() function requires explainers created with specified 'data' parameter.")
   if (is.null(explainer$y)) stop("The variable_importance() function requires explainers created with specified 'y' parameter.")