+++ R/variable_response.R
@@ -58,10 +58,17 @@
 variable_response <- function(explainer, variable, type = "pdp", trans = explainer$link, ...) {
   # Deprecated
   if (type == "pdp") {
-    .Deprecated("ingredients::partial_dependency()", package = "ingredients", msg = "Please note that 'variable_response()' is now deprecated, it is better to use 'ingredients::partial_dependency()' instead.\nFind examples and detailed introduction at: https://pbiecek.github.io/PM_VEE/partialDependenceProfiles.html")
+    if (!exists("message_partial_dependency", envir = .DALEX.env)) {
+      .DALEX.env$message_partial_dependency = TRUE
+      .Deprecated("ingredients::partial_dependency()", package = "ingredients", msg = "Please note that 'variable_response()' is now deprecated, it is better to use 'ingredients::partial_dependency()' instead.\nFind examples and detailed introduction at: https://pbiecek.github.io/PM_VEE/partialDependenceProfiles.html")
+    }
   } else {
-    .Deprecated("ingredients::accumulated_dependency()", package = "ingredients", msg = "Please note that 'variable_response()' is now deprecated, it is better to use 'ingredients::accumulated_dependency()' instead.\nFind examples and detailed introduction at: https://pbiecek.github.io/PM_VEE/accumulatedLocalProfiles.html")
+    if (!exists("message_accumulated_dependency", envir = .DALEX.env)) {
+      .DALEX.env$message_accumulated_dependency = TRUE
+      .Deprecated("ingredients::accumulated_dependency()", package = "ingredients", msg = "Please note that 'variable_response()' is now deprecated, it is better to use 'ingredients::accumulated_dependency()' instead.\nFind examples and detailed introduction at: https://pbiecek.github.io/PM_VEE/accumulatedLocalProfiles.html")
+    }
   }
+
 
   if (!("explainer" %in% class(explainer))) stop("The variable_response() function requires an object created with explain() function.")
   if (is.null(explainer$data)) stop("The variable_response() function requires explainers created with specified 'data' parameter.")